pyproject.toml
[tool.poetry]
name = "clay"
version = "0.1.0"
description = ""
authors = ["Alexander Smirnov <alex@asmirnov.xyz>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
einops = "0.7.0"
loguru = "^0.7.2"
pydantic-settings = "^2.3.4"
python-box = "^7.2.0"
stackstac = "0.5.0"
pystac-client = "^0.8.2"
geopandas = "^1.0.0"
kafka-python-ng = "^2.2.2"
matplotlib = "^3.9.0"
rioxarray = "^0.15.7"
xarray = "^2024.6.0"
scikit-learn = "^1.5.1"
xgboost = "^2.1.0"
# malevich = {url = "https://github.com/MalevichAI/malevich/archive/dev/unstable.zip"}
huggingface-hub = "^0.23.5"
# vit-pytorch = "^1.7.0"


[tool.poetry.group.app.dependencies]
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gunicorn = "^22.0.0"
supabase = "^2.5.1"

[tool.poetry.group.torch.dependencies]
torch = "2.3.1"
torchvision = "0.18.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
traitlets = "5.9.0"
notebook = "6.4.12"
jupyter = "^1.0.0"
jupyter-contrib-nbextensions = "^0.7.0"
jupyterlab = "^4.2.3"

[tool.poetry.group.scripts.dependencies]
fire = "^0.6.0"
pygeohash = "^1.2.0"
qdrant-client = "^1.10.0"
opencv-python = "^4.10.0.84"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
